Minor refactoring and renaming to ensure things do what they say they're doing (and vice versa)
